fosdem: Include day in track name
In some places of the code we make the assumption that the first event
in a track is representative of all events within that track in regards
to day and location. This is mostly true, but not true for some tracks
like "Keynotes" and as such, the day and room of a talk may not
necessarily be correct. By including the day in the name of the keynote
like "Day 1: Keynotes", we aovid this issue.

This also makes it easier to later filter events based on which day of
the conference it's on.